v2.1: 代码质量全面优化
- 安全: SHA-256 → PBKDF2WithHmacSHA256 密码哈希 (10万次迭代)
- 安全: API 登录/注册响应不再返回密码字段
- 架构: AppContext 改用工厂方法,消除 this-escape 警告
- 配置: 新增 AppConfig 支持 -Dserver.port/-Ddata.dir 外部化配置
- 前端: 修复 7 个 Vue 文件中静默吞异常的问题
- 前端: Dashboard 列表 key 改用 recordId 保证唯一性
- 工程: 完善 .gitignore, 新增 MIT LICENSE
- 文档: 重写 README (配置参数表/API分组/安全说明)
- 修复: BookServiceTest 补回误删的 import
- 清理: 删除废弃的 Swing UI 代码

+++ lib/com/lib/demo/entity/BorrowRecord.java
@@ -22,7 +22,7 @@ public class BorrowRecord implements Serializable {
     private LocalDate borrowDate;
     private LocalDate dueDate;
     private LocalDate returnDate;
-    private double fineAmount;
+    private long fineAmount;  // 单位: 分 (50 = ¥0.50)
     private Status status;
     private int renewCount;
 
@@ -43,8 +43,8 @@ public class BorrowRecord implements Serializable {
     public void setDueDate(LocalDate dueDate) { this.dueDate = dueDate; }
     public LocalDate getReturnDate() { return returnDate; }
     public void setReturnDate(LocalDate returnDate) { this.returnDate = returnDate; }
-    public double getFineAmount() { return fineAmount; }
-    public void setFineAmount(double fineAmount) { this.fineAmount = fineAmount; }
+    public long getFineAmount() { return fineAmount; }
+    public void setFineAmount(long fineAmount) { this.fineAmount = fineAmount; }
     public Status getStatus() { return status; }
     public void setStatus(Status status) { this.status = status; }
     public int getRenewCount() { return renewCount; }
